+++ src/id_translation/_tasks/__init__.py
@@ -1,7 +1,9 @@
 from ._map import MappingTask
+from ._names import NamesTask
 from ._translate import TranslationTask
 
 __all__ = [
     "MappingTask",
+    "NamesTask",
     "TranslationTask",
 ]